Refactor amount type in sql db from REAL to BIGINT

+++ in-ou/budget_manager.py
@@ -19,7 +19,7 @@ class BudgetManager:
             CREATE TABLE IF NOT EXISTS budget (
                 id INTEGER PRIMARY KEY AUTOINCREMENT,
                 entry_type TEXT NOT NULL,
-                amount REAL NOT NULL,
+                amount BIGINT NOT NULL,
                 description TEXT,
                 category TEXT,
                 date TEXT NOT NULL
